requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pytest==8.3.4
httpx==0.28.1
PyYAML==6.0.2
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
Pillow==11.0.0
reportlab==4.2.5
